feat: made more stl competible, but it still needs improvements, like allocator template, iterator, stl methods
