+++ src/CMakeLists.txt
@@ -1,5 +1,7 @@
 cmake_minimum_required(VERSION 3.29)
 
 
-add_library(fibHeap
-        STATIC FibHeap.cpp)
+add_library(
+        fibHeap
+        INTERFACE FibHeap.h
+)
